@@ -1,7 +1,7 @@
 class Template < ActiveRecord::Base
   include ActiveModel::ForbiddenAttributesProtection
 
-  has_many :screens, :dependent => :restrict
+  has_many :screens, :dependent => :restrict_with_exception
   has_many :media, :as => :attachable, :dependent => :destroy
   has_many :positions, :dependent => :destroy
   
