@@ -19,15 +19,15 @@ class User < ActiveRecord::Base
   has_many :submissions, :foreign_key => "moderator_id"
   has_many :memberships, :dependent => :destroy
   has_many :groups, :through => :memberships
-  has_many :screens, :as => :owner, :dependent => :restrict
+  has_many :screens, :as => :owner, :dependent => :restrict_with_exception
 
-  has_many :groups, :through => :memberships, :conditions => ["memberships.level > ?", Membership::LEVELS[:pending]]
-  has_many :leading_groups, :through => :memberships, :source => :group, :conditions => {"memberships.level" => Membership::LEVELS[:leader]}
+  has_many :groups, -> { where "memberships.level > ?", Membership::LEVELS[:pending]}, :through => :memberships
+  has_many :leading_groups, -> { where "memberships.level" => Membership::LEVELS[:leader]}, :through => :memberships, :source => :group
 
   # Validations
   validates :first_name, :presence => true
   
-  scope :admin, where(:is_admin => true)
+  scope :admin, -> { where :is_admin => true }
 
   def auto_confirm
     # set as confirmed if we are not confirming user accounts so that if that is ever turned on,
