@@ -11,7 +11,7 @@ if ActiveRecord::Base.connection.table_exists? 'concerto_configs'
       def config.api_key
         if ConcertoConfig[:send_errors] == true
           return '52adf2979c2ab87c634612bef9deaaf2'
-        else 
+        else
           return nil
         end
       end
